Decision Tree Model
This is Decision tree model for machin learsning to classify and predict whether or not the borrower paid back their loan in full.

Data is about Lending Club whcih connects people who need money (borrowers) with people who have money (investors). Hopefully, as an investor you would want to invest in people who showed a profile of having a high probability of paying you back. I am trying to create a model that will help to predict this.